jit: Break out architecture-dependent sources to a sub directory
