+++ erts/emulator/beam/jit/x86/beam_asm_global.cpp
